+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.28.0 **UNRELEASED**
 
+- Added support for publishing through the new [Central Portal](https://central.sonatype.com). To use
+  this use the `CENTRAL_PORTAL` option when specifying the Sonatype host.
 - For Kotlin Multiplatform the main plugin will now automatically publish the
   `release` variant if the project has an Android target.
 - Support specifying the Android variants to publish in `KotlinMultiplatform(...)`.